refactor(enums): relocate Direction/Gender/EquipmentSlot into DALib.Enums
Complete the DALib.Enums consolidation begun in the prior commit. Move the
three protocol enums that already existed elsewhere into the new namespace so
all wire enums live in one place:

- Direction, Gender: out of DALib/Definitions/Enums.cs (file-format enums stay)
- EquipmentSlot: out of DALib/Networking/Packets/Server/ (git-tracked rename)

All three move to namespace DALib.Enums. Consumers retyped: 11 library files
and 10 test files (using DALib.Definitions -> using DALib.Enums where Definitions
became unused; added using DALib.Enums where EquipmentSlot was previously in
the consumer's own namespace). Cross-namespace XML crefs fully qualified
(EquipmentSlot -> Add/RemoveEquipmentPacket; CreateCharFinalizePacket -> Gender).

The Wire/ files matching \bDirection\b reference IPacket.Direction (the packet
direction property), not the movement enum, and are untouched; UserAppearance
and DrawObjects carry Direction as a raw byte and take no enum import.

DALib builds clean (0 errors, 0 warnings); 866/866 tests pass.
