+++ DALib.Tests/Networking/Packets/Server/AddEquipmentPacketTests.cs
@@ -1,3 +1,4 @@
+using DALib.Enums;
 using DALib.Networking.Crypto;
 using DALib.Networking.Packets.Server;
 using DALib.Networking.Wire;
