+++ DALib/Networking/Packets/Server/DisplayUserPacket.cs
@@ -1,5 +1,5 @@
 using System;
-using DALib.Definitions;
+using DALib.Enums;
 using DALib.Networking.Wire;
 
 namespace DALib.Networking.Packets.Server;
